update: move hook into separate file and fix styles
